%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.seq-package}
\alias{aroma.seq-package}
\alias{aroma.seq}
\docType{package}

\title{Package aroma.seq}


 \description{
   [STUB ONLY/DON'T INSTALL] Methods for performing common sequence analysis.  Most of the actual analysis is performed by external software, which this package provides transparent wrappers for.
 }

 \section{Requirements}{
   TBA.
 }

 \section{External software requirements}{
   The following external tools utilized by this package:
   \itemize{
     \item bowtie2: \url{http://bowtie-bio.sourceforge.net/bowtie2/}
     \item BWA: \url{http://bio-bwa.sourceforge.net/}
     \item samtools: \url{http://samtools.sourceforge.net/}
   }

   Some of the external tools are available for Windows from [1].
 }

 \section{Installation and updates}{
   TBA.
 }

 \section{To get started}{
   TBA.
 }

 \section{How to cite this package}{
  Bengtsson H, Olshen A and Tokuyasu T (2013).
\emph{aroma.seq: Sequence analysis in the Aroma Framework}.
R package version 0.2.10.
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.

   The development code of the packages is under a private licence
   (where applicable) and patches sent to the author fall under the
   latter license, but will be, if incorporated, released under the
   "release" license above.
 }

 \references{
   [1] \emph{BOW - Bioinformatics On Windows},
       \url{http://bow.codeplex.com/}.\cr
 }

 \author{Henrik Bengtsson}
\keyword{package}
