%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.AromaSeq.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{AromaSeq}
\docType{class}
\alias{AromaSeq}


 \title{The AromaSeq Package class}

 \description{
  Package:  aroma.seq \cr
\bold{Class AromaSeq}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.oo]{Package}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{\link[aroma.core]{AromaPackage}}\cr
\code{~~~~~~~~~~~~|}\cr
\code{~~~~~~~~~~~~+--}\code{AromaSeq}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{AromaSeq}\cr
extends \link[aroma.core]{AromaPackage}\cr


 }

 \usage{
AromaSeq(...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[aroma.seq:capabilitiesOf.AromaSeq]{capabilitiesOf}} \tab Checks which tools are supported.\cr
 \tab \code{isCapableOf} \tab  -\cr
 \tab \code{setupTests} \tab  -\cr
}


 \bold{Methods inherited from AromaPackage}:\cr
fixSearchPath, fixSearchPathInternal, patch

 \bold{Methods inherited from Package}:\cr
as.character, getAuthor, getBundle, getBundlePackages, getChangeLog, getClasses, getContents, getDataPath, getDate, getDescription, getDescriptionFile, getDocPath, getEnvironment, getExamplePath, getHistory, getHowToCite, getLicense, getMaintainer, getName, getNews, getPath, getPosition, getTitle, getUrl, getVersion, isLoaded, isOlderThan, ll, load, showChangeLog, showContents, showDescriptionFile, showHistory, showHowToCite, showNews, startupMessage, unload, update, browseRsp

 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save, findRspReportTemplate, asThis


 }

 \author{Henrik Bengtsson}
\keyword{classes}
