%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Bowtie2Alignment.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Bowtie2Aligment}
\docType{class}
\alias{Bowtie2Aligment}


 \title{The Bowtie2Aligment class}

 \description{
  Package:  aroma.seq \cr
\bold{Class Bowtie2Aligment}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Bowtie2Aligment}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Bowtie2Aligment}\cr
extends \link[R.oo]{Object}\cr



  ...
 }

 \usage{Bowtie2Aligment(ds=NULL, reference=NULL, tags="*", ...)}

 \arguments{
  \item{ds}{An \code{\link{FastqDataSet}}.}
  \item{reference}{An \code{\link{FastaReferenceFile}}.} of
  \item{tags}{Additional tags for the output data sets.}
  \item{...}{Additional arguments passed to the aligner.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{getFullName} \tab  -\cr
 \tab \code{getInputDataSet} \tab  -\cr
 \tab \code{getName} \tab  -\cr
 \tab \code{getOutputDataSet} \tab  -\cr
 \tab \code{getPath} \tab  -\cr
 \tab \code{getReferenceFile} \tab  -\cr
 \tab \code{getRootPath} \tab  -\cr
 \tab \code{getTags} \tab  -\cr
 \tab \code{nbrOfFiles} \tab  -\cr
 \tab \code{process} \tab  -\cr
 \tab \code{setTags} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


 }

 \author{Henrik Bengtsson and Pierre Neuvial}
\keyword{classes}
