%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cufflinks.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cufflinks}
\alias{cufflinks.default}
\alias{cufflinks}

\title{Calls cufflinks on input bam file(s)}

\description{
 Calls cufflinks on input bam file(s).
}

\usage{\method{cufflinks}{default}(commandName="cufflinks", bams=NULL, cufflinksOptions, ..., verbose=FALSE)}  - level2

\arguments{
}

\examples{\dontrun{
}}

\author{TBA}

