%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  findFastqDump.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findFastqDump}
\alias{findFastqDump}

\alias{}

\title{Locates the FastqDump executable}

\description{
 Locates the FastqDump executable on the current system.
}

\usage{
findFastqDump(commandName="fastq-dump", ...)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\details{
}

\author{Taku Tokuyasu}


\keyword{internal}
