%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.FastaReferenceFile}
\alias{as.character.FastqDataFile}
\alias{getCommonSeqWidth}
\alias{getCommonSeqWidth.FastqDataFile}
\alias{getGeometry}
\alias{getGeometry.FastqDataFile}
\alias{getSeqLengths}
\alias{getSeqLengths.FastaReferenceFile}
\alias{getSeqNames}
\alias{getSeqNames.FastaReferenceFile}
\alias{getTotalSeqLengths}
\alias{getTotalSeqLengths.FastaReferenceFile}
\alias{nbrOfSeqs}
\alias{nbrOfSeqs.FastaReferenceFile}
\alias{nbrOfSeqs.FastqDataFile}
\alias{readGeometry}
\alias{readGeometry.FastqDataFile}
\alias{readSeqLengths}
\alias{readSeqLengths.FastaReferenceFile}
\alias{validate.FastqDataSet}
\alias{as.character.BamDataFile}
\alias{getFlavor}
\alias{getHeader.BamDataFile}
\alias{getInputDataSet}
\alias{getOutputDataSet}
\alias{getReferenceFile}
\alias{getRootPath}
\alias{getTargetLengths}
\alias{getTargetLengths.BamDataFile}
\alias{getTargetNames}
\alias{getTargetNames.BamDataFile}
\alias{getTargets}
\alias{getTargets.BamDataFile}
\alias{getTotalTargetLength}
\alias{getTotalTargetLength.BamDataFile}
\alias{hasIndex}
\alias{hasIndex.BamDataFile}
\alias{isSorted}
\alias{isSorted.BamDataFile}
\alias{nbrOfSeqs.BamDataFile}
\alias{nbrOfTargets}
\alias{nbrOfTargets.BamDataFile}
\alias{process}
\alias{processOne}
\alias{readHeader}
\alias{readHeader.BamDataFile}
\alias{sort.BamDataFile}
\alias{validate.BamDataSet}
\alias{bowtie2}
\alias{bowtie2.default}
\alias{as.character.Bowtie2Alignment}
\alias{bDebugSystemBowtie2Build}
\alias{bowtie2Build}
\alias{bowtie2Build.default}
\alias{getAsteriskTags.Bowtie2Alignment}
\alias{getFlavor.Bowtie2Alignment}
\alias{getFullName.Bowtie2Alignment}
\alias{getInputDataSet.Bowtie2Alignment}
\alias{getName.Bowtie2Alignment}
\alias{getOutputDataSet.Bowtie2Alignment}
\alias{getPath.Bowtie2Alignment}
\alias{getReferenceFile.Bowtie2Alignment}
\alias{getRootPath.Bowtie2Alignment}
\alias{getTags.Bowtie2Alignment}
\alias{nbrOfFiles.Bowtie2Alignment}
\alias{process.Bowtie2Alignment}
\alias{processOne.Bowtie2Alignment}
\alias{setTags.Bowtie2Alignment}
\alias{systemBowtie2}
\alias{systemBowtie2.default}
\alias{systemBowtie2Build}
\alias{systemBowtie2Build.default}
\alias{testSystemBowtie2Build}
\alias{isCapableOf}
\alias{buildBwaIndexSet}
\alias{as.character.BwaAlignment}
\alias{as.character.BwaIndexFile}
\alias{as.character.BwaIndexSet}
\alias{as.character.SamDataFile}
\alias{getAsteriskTags.BwaAlignment}
\alias{getFlavor.BwaAlignment}
\alias{getFullName.BwaAlignment}
\alias{getHeader.SamDataFile}
\alias{getIndexPrefix}
\alias{getIndexPrefix.BwaIndexFile}
\alias{getIndexPrefix.BwaIndexSet}
\alias{getIndexSet}
\alias{getIndexSet.BwaAlignment}
\alias{getInputDataSet.BwaAlignment}
\alias{getName.BwaAlignment}
\alias{getOutputDataSet.BwaAlignment}
\alias{getPath.BwaAlignment}
\alias{getRootPath.BwaAlignment}
\alias{getTags.BwaAlignment}
\alias{getTargetLengths.SamDataFile}
\alias{getTargetNames.SamDataFile}
\alias{getTargets.SamDataFile}
\alias{getTotalTargetLength.SamDataFile}
\alias{hasIndex.SamDataFile}
\alias{isComplete}
\alias{isComplete.BwaIndexSet}
\alias{isSorted.SamDataFile}
\alias{nbrOfFiles.BwaAlignment}
\alias{nbrOfSeqs.SamDataFile}
\alias{nbrOfTargets.SamDataFile}
\alias{readHeader.SamDataFile}
\alias{setTags.BwaAlignment}
\alias{sort.SamDataFile}
\alias{validate.SamDataSet}
\alias{byPrefix}
\alias{byPrefix.BwaIndexSet}
\alias{getDepth.BamDataSet}
\alias{getDepth.FastqDataSet}
\alias{getDepth.SamDataSet}
\alias{byPath.SamDataSet}
\alias{convertToBamDataFile}
\alias{convertToBamDataFile.SamDataFile}
\alias{convertToBamDataSet}
\alias{byPath.BamDataSet}
\alias{getOptionalArguments.BwaAlignment}
\alias{capabilitiesOf}
\alias{setupTests}
\alias{setupTests.AromaSeq}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
