%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  samtoolsFaidx.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{samtoolsFaidx}
\alias{samtoolsFaidx.default}
\alias{samtoolsFaidx}

\title{Calls the samtools 'faidx' command}

\description{
 Calls the samtools 'faidx' command.
}

\usage{
\method{samtoolsFaidx}{default}(pathname, ..., verbose=FALSE)
}

\arguments{
  \item{pathname}{A reference fasta file.}
  \item{...}{Additional arguments specifying samtools 'faidx' switches
    passed to \code{\link{systemSamtools}}().}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\author{Taku Tokuyasu}


\keyword{internal}
