%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SequenceContigsInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{SequenceContigsInterface}
\docType{class}
\alias{SequenceContigsInterface}


\title{The SequenceContigsInterface class}

\description{
 Package:  aroma.seq \cr
\bold{Class SequenceContigsInterface}\cr

\code{\link[R.oo]{Interface}}\cr
\code{~~|}\cr
\code{~~+--}\emph{\code{SequenceContigsInterface}}\cr

\bold{Directly known subclasses:}\cr
\link[aroma.seq]{BamDataFile}, \link[aroma.seq]{Bowtie2IndexSet}, \emph{\link[aroma.seq]{FastaReferenceFile}}, \emph{\link[aroma.seq]{FastaReferenceIndexFile}}, \emph{\link[aroma.seq]{GcBaseFile}}, \emph{\link[aroma.seq]{GtfDataFile}}\cr

public abstract class \bold{SequenceContigsInterface}\cr
extends \link[R.oo]{Interface}\cr



 An \code{\link[R.oo]{Interface}} for methods for with sequence contigs.
}

\usage{
SequenceContigsInterface(...)
}

\arguments{
  \item{...}{Not used.}
}

\section{Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{getSeqLengths} \tab  -\cr
 \tab \code{getSeqNames} \tab  -\cr
 \tab \code{getTotalSeqLength} \tab  -\cr
 \tab \code{hasSeqLengths} \tab  -\cr
 \tab \code{nbrOfSeqs} \tab  -\cr
}


 \bold{Methods inherited from Interface}:\cr
extend, print, uses


}

\author{Henrik Bengtsson}
\keyword{classes}
