%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  systemBWA.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{systemBWA}
\alias{systemBWA.default}
\alias{systemBWA}

 \title{Calls the BWA executable}

 \description{
  Calls the BWA executable.
 }

 \usage{\method{systemBWA}{default}(command, ..., .fake=FALSE, verbose=FALSE)}

 \arguments{
   \item{command}{A \code{\link[base]{character}} string specifying the BWA command.}
   \item{...}{Additional arguments specifying BWA command line switches.}
   \item{.fake}{If \code{\link[base:logical]{TRUE}}, the executable is not called.}
   \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

