%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  systemFastQC.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{systemFastQC}
\alias{systemFastQC.default}
\alias{systemFastQC}

\title{Calls the FastQC executable}

\description{
 Calls the FastQC executable.
}

\usage{
\method{systemFastQC}{default}(..., verbose=FALSE)
}

\arguments{
  \item{...}{Additional arguments specifying FastQC command line switches.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\author{Henrik Bengtsson}


\keyword{internal}
