%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  systemJava.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{systemJavaJar}
\alias{systemJavaJar.default}
\alias{systemJavaJar}

 \title{Calls the Java jar executable}

 \description{
  Calls the Java jar executable.
 }

 \usage{\method{systemJavaJar}{default}(pathname, ...)}

 \arguments{
   \item{pathname}{The JAR file.}
   \item{...}{Additional arguments specifying Java command line switches.}
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

