%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  systemTopHatBuild.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{systemTopHatBuild}
\alias{systemTopHatBuild.default}
\alias{systemTopHatBuild}

\title{Calls the TopHat executable, specifically to build a transcriptome index}

\description{
 Calls the TopHat executable, specifically to build a transcriptome index.
}

\usage{
\method{systemTopHatBuild}{default}(..., system2Args, .fake=FALSE, verbose=FALSE)
}

\arguments{
  \item{...}{Arguments specifying TopHat command line switches.}
  \item{.fake}{If \code{\link[base:logical]{TRUE}}, the executable is not called.}
  \item{system2Args}{Arguments to be passed to system2}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}


\author{Taku Tokuyasu}


\keyword{internal}
