%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  tophat.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{tophat}
\alias{tophat.default}
\alias{tophat}

\title{Calls tophat on input reads}

\description{
 Calls tophat on input reads.
}

\usage{\method{tophat}{default}(commandName="tophat", bowtieRefIndexPrefix, reads1=NA, reads2=NA, tophatOptions, ..., verbose=FALSE)}

\arguments{
  \item{...}{...}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

