%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.seq-package}
\alias{aroma.seq-package}
\alias{aroma.seq}
\docType{package}

\title{Package aroma.seq}


 \description{
   [STUB ONLY/DON'T INSTALL] Methods for performing common sequence analysis.  Most of the actual analysis is performed by external software, which this package provides transparent wrappers for.
 }

 \section{Requirements}{
   TBA.
 }

 \section{Installation and updates}{
   TBA.
 }

 \section{To get started}{
   TBA.
 }

 \section{How to cite this package}{
   TBA.
 }

 \author{
   Henrik Bengtsson et al.
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.

   The development code of the packages is under a private licence
   (where applicable) and patches sent to the author fall under the
   latter license, but will be, if incorporated, released under the
   "release" license above.
 }

 \references{
   TBA
 }
\keyword{package}
