%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cuffdiff.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cuffdiff}
\alias{cuffdiff.default}
\alias{cuffdiff}

\title{Calls cuffdiff on input bam file(s)}

\description{
 Calls cuffdiff on input bam file(s).
}

\usage{
\method{cuffdiff}{default}(transcriptsGtf, bams=NULL, optionsVec=NULL, ..., system2ArgsList=NULL, commandName="cuffdiff", verbose=FALSE)
}

\arguments{
}

\examples{\dontrun{
}}

\author{Henrik Bengtsson}

