%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cuffdiff.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cuffdiff}
\alias{cuffdiff.default}
\alias{cuffdiff}

\title{Calls cuffdiff on input bam file(s)}

\description{
 Calls cuffdiff on input bam file(s).
}

\usage{\method{cuffdiff}{default}(commandName="cuffdiff", bams=NULL, cuffdiffOptions, ..., verbose=FALSE)}  - level2

\arguments{
}

\examples{\dontrun{
}}

\author{TBA}

