%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cufflinks.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cufflinks}
\alias{cufflinks.default}
\alias{cufflinks}

\title{Calls cufflinks on input bam file(s)}

\description{
 Calls cufflinks on input bam file(s).
}

\usage{
\method{cufflinks}{default}(bams=NULL, optionsVec, ..., commandName="cufflinks", verbose=FALSE)
}

\arguments{
  \item{bams}{Vector of pathnames (.sam or .bam)}
  \item{optionsVec}{Vector of named options}
  \item{...}{...}
  \item{commandName}{Name of executable}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\author{Taku Tokuyasu}


\keyword{internal}
