%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cufflinks.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cufflinks}
\alias{cufflinks.default}
\alias{cufflinks}

\title{Calls cufflinks on input bam file(s)}

\description{
 Calls cufflinks on input bam file(s).
}

\usage{
\method{cufflinks}{default}(bams=NULL, optionsVec, ..., commandName="cufflinks", verbose=FALSE)
}

\arguments{
}

\examples{\dontrun{
}}

\author{Henrik Bengtsson}

