%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  fastQC.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fastQC}
\alias{fastQC.default}
\alias{fastQC}

\title{Calls the FastQC executable}

\description{
 Calls the FastQC executable.
}

\usage{
\method{fastQC}{default}(..., pathnames=character(0L), outPath=NULL, verbose=FALSE)
}

\arguments{
  \item{pathnames}{Zero or more input pathnames.}
  \item{...}{Additional arguments passed to \code{\link{systemFastQC}}().}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\author{Henrik Bengtsson}


\keyword{internal}
