%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  gatk.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{gatk}
\alias{gatk}

\alias{gatk.default}

\title{Calls the GATK executable}

\description{
 Calls the GATK executable.
}

\usage{
 \method{gatk}{default}(..., analysisType=NULL, inputFile=NULL, referenceSequence=NULL, interval=NULL,
  outPath="gatkData/", verbose=FALSE)
}

\arguments{
  \item{...}{(Not used)}.
  \item{outPath}{Directory where result files are written.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\author{Henrik Bengtsson}

\references{
 [1] GATK: The Genome Analysis Toolkit,
     Broad Institute, 2014.
     \url{http://www.broadinstitute.org/gatk/}
}


\keyword{methods}
\keyword{internal}
