%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.FastaReferenceFile}
\alias{as.character.FastqDataFile}
\alias{getCommonSeqWidth}
\alias{getCommonSeqWidth.FastqDataFile}
\alias{getGeometry}
\alias{getGeometry.FastqDataFile}
\alias{getSeqLengths}
\alias{getSeqLengths.FastaReferenceFile}
\alias{getSeqNames}
\alias{getSeqNames.FastaReferenceFile}
\alias{getTotalSeqLengths}
\alias{getTotalSeqLengths.FastaReferenceFile}
\alias{nbrOfSeqs}
\alias{nbrOfSeqs.FastaReferenceFile}
\alias{nbrOfSeqs.FastqDataFile}
\alias{readGeometry}
\alias{readGeometry.FastqDataFile}
\alias{readSeqLengths}
\alias{readSeqLengths.FastaReferenceFile}
\alias{validate.FastqDataSet}
\alias{as.character.BamDataFile}
\alias{getFlavor}
\alias{getHeader.BamDataFile}
\alias{getInputDataSet}
\alias{getOutputDataSet}
\alias{getReferenceFile}
\alias{getRootPath}
\alias{getTargetLengths}
\alias{getTargetLengths.BamDataFile}
\alias{getTargetNames}
\alias{getTargetNames.BamDataFile}
\alias{getTargets}
\alias{getTargets.BamDataFile}
\alias{getTotalTargetLength}
\alias{getTotalTargetLength.BamDataFile}
\alias{hasIndex}
\alias{hasIndex.BamDataFile}
\alias{isSorted}
\alias{isSorted.BamDataFile}
\alias{nbrOfSeqs.BamDataFile}
\alias{nbrOfTargets}
\alias{nbrOfTargets.BamDataFile}
\alias{process}
\alias{processOne}
\alias{readHeader}
\alias{readHeader.BamDataFile}
\alias{sort.BamDataFile}
\alias{validate.BamDataSet}
\alias{bowtie2}
\alias{bowtie2.default}
\alias{as.character.Bowtie2Alignment}
\alias{bDebugSystemBowtie2Build}
\alias{bowtie2Build}
\alias{bowtie2Build.default}
\alias{getAsteriskTags.Bowtie2Alignment}
\alias{getFlavor.Bowtie2Alignment}
\alias{getFullName.Bowtie2Alignment}
\alias{getInputDataSet.Bowtie2Alignment}
\alias{getName.Bowtie2Alignment}
\alias{getOutputDataSet.Bowtie2Alignment}
\alias{getPath.Bowtie2Alignment}
\alias{getReferenceFile.Bowtie2Alignment}
\alias{getRootPath.Bowtie2Alignment}
\alias{getTags.Bowtie2Alignment}
\alias{nbrOfFiles.Bowtie2Alignment}
\alias{processOne.Bowtie2Alignment}
\alias{setTags.Bowtie2Alignment}
\alias{systemBowtie2}
\alias{systemBowtie2.default}
\alias{systemBowtie2Build}
\alias{systemBowtie2Build.default}
\alias{testSystemBowtie2Build}
\alias{isCapableOf}
\alias{buildBwaIndexSet}
\alias{as.character.BwaAlignment}
\alias{as.character.BwaIndexFile}
\alias{as.character.BwaIndexSet}
\alias{as.character.SamDataFile}
\alias{getAsteriskTags.BwaAlignment}
\alias{getFlavor.BwaAlignment}
\alias{getFullName.BwaAlignment}
\alias{getHeader.SamDataFile}
\alias{getIndexPrefix}
\alias{getIndexPrefix.BwaIndexFile}
\alias{getIndexPrefix.BwaIndexSet}
\alias{getIndexSet}
\alias{getIndexSet.BwaAlignment}
\alias{getInputDataSet.BwaAlignment}
\alias{getName.BwaAlignment}
\alias{getOutputDataSet.BwaAlignment}
\alias{getPath.BwaAlignment}
\alias{getRootPath.BwaAlignment}
\alias{getTags.BwaAlignment}
\alias{getTargetLengths.SamDataFile}
\alias{getTargetNames.SamDataFile}
\alias{getTargets.SamDataFile}
\alias{getTotalTargetLength.SamDataFile}
\alias{hasIndex.SamDataFile}
\alias{isComplete}
\alias{isComplete.BwaIndexSet}
\alias{isSorted.SamDataFile}
\alias{nbrOfFiles.BwaAlignment}
\alias{nbrOfSeqs.SamDataFile}
\alias{nbrOfTargets.SamDataFile}
\alias{readHeader.SamDataFile}
\alias{setTags.BwaAlignment}
\alias{sort.SamDataFile}
\alias{validate.SamDataSet}
\alias{byPrefix}
\alias{byPrefix.BwaIndexSet}
\alias{getDepth.BamDataSet}
\alias{getDepth.FastqDataSet}
\alias{getDepth.SamDataSet}
\alias{byPath.SamDataSet}
\alias{convertToBamDataFile}
\alias{convertToBamDataFile.SamDataFile}
\alias{convertToBamDataSet}
\alias{byPath.BamDataSet}
\alias{getOptionalArguments.BwaAlignment}
\alias{capabilitiesOf}
\alias{setupTests}
\alias{setupTests.AromaSeq}
\alias{as.character.AbstractIndexFile}
\alias{as.character.AbstractIndexSet}
\alias{bowtie2IndexPrefix}
\alias{bowtie2IndexPrefix.default}
\alias{buildBowtie2IndexSet}
\alias{buildBowtie2IndexSet.FastaReferenceFile}
\alias{byPrefix.AbstractIndexSet}
\alias{getIndexPrefix.AbstractIndexFile}
\alias{getIndexPrefix.AbstractIndexSet}
\alias{isComplete.AbstractIndexSet}
\alias{isComplete.Bowtie2IndexSet}
\alias{getIndexSet.Bowtie2Alignment}
\alias{getOptionalArguments.Bowtie2Alignment}
\alias{getSequenceNames}
\alias{getSequenceNames.Bowtie2IndexSet}
\alias{getSummary}
\alias{getSummary.Bowtie2IndexSet}
\alias{buildIndex}
\alias{buildIndex.BamDataFile}
\alias{getReadGroups}
\alias{getReadGroups.BamDataFile}
\alias{replaceAllReadGroups}
\alias{replaceAllReadGroups.BamDataFile}
\alias{as.character.SamReadGroup}
\alias{asBwaString}
\alias{asBwaString.SamReadGroup}
\alias{asSamList}
\alias{asSamList.SamReadGroup}
\alias{asString}
\alias{asString.SamReadGroup}
\alias{assignBy}
\alias{assignBy.SamReadGroup}
\alias{byScanBamHeader}
\alias{byScanBamHeader.SamReadGroup}
\alias{getParameters.Bowtie2Alignment}
\alias{getParameters.BwaAlignment}
\alias{hasDescription}
\alias{hasDescription.SamReadGroup}
\alias{hasLibrary}
\alias{hasLibrary.SamReadGroup}
\alias{hasPlatform}
\alias{hasPlatform.SamReadGroup}
\alias{hasPlatformUnit}
\alias{hasPlatformUnit.SamReadGroup}
\alias{hasRunDate}
\alias{hasRunDate.SamReadGroup}
\alias{hasSample}
\alias{hasSample.SamReadGroup}
\alias{hasSequencingCenter}
\alias{hasSequencingCenter.SamReadGroup}
\alias{validate.SamReadGroup}
\alias{extractSamReadGroup}
\alias{extractSamReadGroup.FastqDataFile}
\alias{extractSamReadGroup.IlluminaFastqDataFile}
\alias{getBarcodeSequence}
\alias{getBarcodeSequence.IlluminaFastqDataFile}
\alias{getFirstSequenceInfo}
\alias{getFirstSequenceInfo.IlluminaFastqDataFile}
\alias{getFlowcellId}
\alias{getFlowcellId.IlluminaFastqDataFile}
\alias{getInstrumentId}
\alias{getInstrumentId.IlluminaFastqDataFile}
\alias{getLane}
\alias{getLane.IlluminaFastqDataFile}
\alias{getPlatform.IlluminaFastqDataFile}
\alias{getPlatformUnit}
\alias{getPlatformUnit.IlluminaFastqDataFile}
\alias{getSampleName.IlluminaFastqDataFile}
\alias{merge.SamReadGroup}
\alias{as.character.AbstractAlignment}
\alias{getAsteriskTags.AbstractAlignment}
\alias{getDefaultSamReadGroup}
\alias{getDefaultSamReadGroup.FastqDataFile}
\alias{getDefaultSamReadGroup.FastqDataSet}
\alias{getDefaultSamReadGroup.IlluminaFastqDataFile}
\alias{getFlavor.AbstractAlignment}
\alias{getFullName.AbstractAlignment}
\alias{getIndexSet.AbstractAlignment}
\alias{getInputDataSet.AbstractAlignment}
\alias{getName.AbstractAlignment}
\alias{getOptionalArguments.AbstractAlignment}
\alias{getOutputDataSet.AbstractAlignment}
\alias{getParameterSets.AbstractAlignment}
\alias{getPath.AbstractAlignment}
\alias{getRootPath.AbstractAlignment}
\alias{getSamReadGroup}
\alias{getSamReadGroup.FastqDataFile}
\alias{getSamReadGroup.FastqDataSet}
\alias{getTags.AbstractAlignment}
\alias{isEmpty}
\alias{isEmpty.SamReadGroup}
\alias{nbrOfFiles.AbstractAlignment}
\alias{process.AbstractAlignment}
\alias{setSamReadGroup}
\alias{setSamReadGroup.FastqDataFile}
\alias{setSamReadGroup.FastqDataSet}
\alias{setTags.AbstractAlignment}
\alias{getReadGroup}
\alias{getReadGroup.BamDataFile}
\alias{hasID}
\alias{hasID.SamReadGroup}
\alias{as.character.PicardDuplicateRemoval}
\alias{getAsteriskTags.PicardDuplicateRemoval}
\alias{getFileVersion}
\alias{getFileVersion.IlluminaFastqDataFile}
\alias{getFlavor.PicardDuplicateRemoval}
\alias{getFullName.PicardDuplicateRemoval}
\alias{getInputDataSet.PicardDuplicateRemoval}
\alias{getName.PicardDuplicateRemoval}
\alias{getOptionalArguments.PicardDuplicateRemoval}
\alias{getOutputDataSet.PicardDuplicateRemoval}
\alias{getParameterSets.PicardDuplicateRemoval}
\alias{getPath.PicardDuplicateRemoval}
\alias{getReadDirection}
\alias{getReadDirection.IlluminaFastqDataFile}
\alias{getRootPath.PicardDuplicateRemoval}
\alias{getTags.PicardDuplicateRemoval}
\alias{nbrOfFiles.PicardDuplicateRemoval}
\alias{process.PicardDuplicateRemoval}
\alias{setTags.PicardDuplicateRemoval}
\alias{as.character.BamIndexDataFile}
\alias{getIndexFile}
\alias{getIndexFile.BamDataFile}
\alias{getIndexPrefix.Bowtie2IndexFile}
\alias{getIndexStats}
\alias{getIndexStats.BamDataFile}
\alias{getReadCounts}
\alias{getReadCounts.BamDataFile}
\alias{nbrOfMappedReads}
\alias{nbrOfMappedReads.BamDataFile}
\alias{nbrOfReads.BamDataFile}
\alias{nbrOfUnmappedReads}
\alias{nbrOfUnmappedReads.BamDataFile}
\alias{extractReadStartPositions}
\alias{extractReadStartPositions.BamDataFile}
\alias{readDataFrame.BamDataFile}
\alias{readReadPositions}
\alias{readReadPositions.BamDataFile}
\alias{binTabulate.MaskedDNAString}
\alias{binTabulate}
\alias{extractRawCopyNumbers.BamDataFile}
\alias{getChipType.BamDataFile}
\alias{getChipType.BamDataSet}
\alias{getExpectedOutputFullnames.TotalCnBinnedCounting}
\alias{getFilenameExtension.BamDataFile}
\alias{getOutputFileExtension}
\alias{getOutputFileExtension.TotalCnBinnedCounting}
\alias{getOutputFileSetClass}
\alias{getOutputFileSetClass.TotalCnBinnedCounting}
\alias{getPlatform.BamDataFile}
\alias{getPlatform.BamDataSet}
\alias{smoothRawCopyNumbers}
\alias{smoothRawCopyNumbers.TotalCnBinnedCounting}
\alias{AromaUncFile}
\alias{AromaUnitNucleotideCountsFile}
\alias{allocateFromUgp}
\alias{allocateFromUgp.AromaUnitNucleotideCountsFile}
\alias{getAromaUncFile}
\alias{getAromaUncFile.AromaUgpFile}
\alias{getColumnNames.AromaUnitNucleotideCountsFile}
\alias{getExtensionPattern.AromaUnitNucleotideCountsFile}
\alias{getFilenameExtension.AromaUnitNucleotideCountsFile}
\alias{importFromBSgenome}
\alias{importFromBSgenome.AromaUnitNucleotideCountsFile}
\alias{as.character.BinnedGcNormalization}
\alias{getAsteriskTags.BinnedGcNormalization}
\alias{getGcContent}
\alias{getGcContent.AromaUnitNucleotideCountsFile}
\alias{getGcContentFile}
\alias{getGcContentFile.BinnedGcNormalization}
\alias{getOutputDataSet0.BinnedGcNormalization}
\alias{getOutputFileClass}
\alias{getOutputFileClass.BinnedGcNormalization}
\alias{getOutputFileExtension.BinnedGcNormalization}
\alias{getOutputFileSetClass.BinnedGcNormalization}
\alias{getOutputFiles.BinnedGcNormalization}
\alias{getParameters.BinnedGcNormalization}
\alias{getPath.BinnedGcNormalization}
\alias{getRootPath.BinnedGcNormalization}
\alias{process.BinnedGcNormalization}
\alias{GatkAlleleCounting}
\alias{anyDuplicated.AromaUgpFile}
\alias{getParameters.GatkAlleleCounting}
\alias{getPath.GatkAlleleCounting}
\alias{getRootPath.GatkAlleleCounting}
\alias{process.GatkAlleleCounting}
\alias{readGatkCountFile}
\alias{readGatkCountFile.GatkAlleleCounting}
\alias{validate.BamDataFile}
\alias{writeBedDataFile}
\alias{writeBedDataFile.AromaUgpFile}
\alias{getDefaultColumnNames.AromaUnitNucleotideCountsFile}
\alias{getCombineBy}
\alias{getCombineBy.GatkAlleleCounting}
\alias{getExpectedOutputFullnames.GatkAlleleCounting}
\alias{getOutputDataSet0.GatkAlleleCounting}
\alias{getParameterSets.BwaAlignment}
\alias{getParameters.PicardDuplicateRemoval}
\alias{FastqQualityReport}
\alias{findRspReportTemplate}
\alias{findRspReportTemplate.Object}
\alias{report}
\alias{report.FastqDataFile}
\alias{report.FastqDataSet}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
