%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.FastaReferenceFile}
\alias{as.character.FastqDataFile}
\alias{getCommonSeqWidth}
\alias{getCommonSeqWidth.FastqDataFile}
\alias{getGeometry}
\alias{getGeometry.FastqDataFile}
\alias{getSeqLengths}
\alias{getSeqLengths.FastaReferenceFile}
\alias{getSeqNames}
\alias{getSeqNames.FastaReferenceFile}
\alias{getTotalSeqLengths}
\alias{getTotalSeqLengths.FastaReferenceFile}
\alias{nbrOfSeqs}
\alias{nbrOfSeqs.FastaReferenceFile}
\alias{nbrOfSeqs.FastqDataFile}
\alias{readGeometry}
\alias{readGeometry.FastqDataFile}
\alias{readSeqLengths}
\alias{readSeqLengths.FastaReferenceFile}
\alias{validate.FastqDataSet}
\alias{as.character.BamDataFile}
\alias{getFlavor}
\alias{getHeader.BamDataFile}
\alias{getInputDataSet}
\alias{getReferenceFile}
\alias{getRootPath}
\alias{getTargetLengths}
\alias{getTargetLengths.BamDataFile}
\alias{getTargetNames}
\alias{getTargetNames.BamDataFile}
\alias{getTargets}
\alias{getTargets.BamDataFile}
\alias{getTotalTargetLength}
\alias{getTotalTargetLength.BamDataFile}
\alias{hasIndex}
\alias{hasIndex.BamDataFile}
\alias{isSorted}
\alias{isSorted.BamDataFile}
\alias{nbrOfSeqs.BamDataFile}
\alias{nbrOfTargets}
\alias{nbrOfTargets.BamDataFile}
\alias{process}
\alias{processOne}
\alias{readHeader}
\alias{readHeader.BamDataFile}
\alias{sort.BamDataFile}
\alias{validate.BamDataSet}
\alias{as.character.Bowtie2Alignment}
\alias{bDebugSystemBowtie2Build}
\alias{getAsteriskTags.Bowtie2Alignment}
\alias{getFlavor.Bowtie2Alignment}
\alias{getFullName.Bowtie2Alignment}
\alias{getInputDataSet.Bowtie2Alignment}
\alias{getName.Bowtie2Alignment}
\alias{getPath.Bowtie2Alignment}
\alias{getReferenceFile.Bowtie2Alignment}
\alias{getRootPath.Bowtie2Alignment}
\alias{getTags.Bowtie2Alignment}
\alias{nbrOfFiles.Bowtie2Alignment}
\alias{processOne.Bowtie2Alignment}
\alias{setTags.Bowtie2Alignment}
\alias{testSystemBowtie2Build}
\alias{isCapableOf}
\alias{buildBwaIndexSet}
\alias{as.character.BwaAlignment}
\alias{as.character.BwaIndexFile}
\alias{as.character.BwaIndexSet}
\alias{as.character.SamDataFile}
\alias{getAsteriskTags.BwaAlignment}
\alias{getFlavor.BwaAlignment}
\alias{getFullName.BwaAlignment}
\alias{getHeader.SamDataFile}
\alias{getIndexPrefix}
\alias{getIndexPrefix.BwaIndexFile}
\alias{getIndexPrefix.BwaIndexSet}
\alias{getIndexSet}
\alias{getIndexSet.BwaAlignment}
\alias{getInputDataSet.BwaAlignment}
\alias{getName.BwaAlignment}
\alias{getPath.BwaAlignment}
\alias{getRootPath.BwaAlignment}
\alias{getTags.BwaAlignment}
\alias{getTargetLengths.SamDataFile}
\alias{getTargetNames.SamDataFile}
\alias{getTargets.SamDataFile}
\alias{getTotalTargetLength.SamDataFile}
\alias{hasIndex.SamDataFile}
\alias{isComplete}
\alias{isComplete.BwaIndexSet}
\alias{isSorted.SamDataFile}
\alias{nbrOfFiles.BwaAlignment}
\alias{nbrOfSeqs.SamDataFile}
\alias{nbrOfTargets.SamDataFile}
\alias{readHeader.SamDataFile}
\alias{setTags.BwaAlignment}
\alias{sort.SamDataFile}
\alias{validate.SamDataSet}
\alias{byPrefix}
\alias{byPrefix.BwaIndexSet}
\alias{getDepth.BamDataSet}
\alias{getDepth.FastqDataSet}
\alias{getDepth.SamDataSet}
\alias{byPath.SamDataSet}
\alias{convertToBamDataFile}
\alias{convertToBamDataFile.SamDataFile}
\alias{convertToBamDataSet}
\alias{byPath.BamDataSet}
\alias{getOptionalArguments.BwaAlignment}
\alias{capabilitiesOf}
\alias{setupTests}
\alias{setupTests.AromaSeq}
\alias{as.character.AbstractIndexFile}
\alias{as.character.AbstractIndexSet}
\alias{bowtie2IndexPrefix}
\alias{bowtie2IndexPrefix.default}
\alias{buildBowtie2IndexSet}
\alias{byPrefix.AbstractIndexSet}
\alias{getIndexPrefix.AbstractIndexFile}
\alias{getIndexPrefix.AbstractIndexSet}
\alias{isComplete.AbstractIndexSet}
\alias{isComplete.Bowtie2IndexSet}
\alias{getIndexSet.Bowtie2Alignment}
\alias{getOptionalArguments.Bowtie2Alignment}
\alias{getSequenceNames}
\alias{getSequenceNames.Bowtie2IndexSet}
\alias{getSummary}
\alias{getSummary.Bowtie2IndexSet}
\alias{buildIndex}
\alias{buildIndex.BamDataFile}
\alias{getReadGroups}
\alias{getReadGroups.BamDataFile}
\alias{replaceAllReadGroups}
\alias{replaceAllReadGroups.BamDataFile}
\alias{as.character.SamReadGroup}
\alias{asBwaString}
\alias{asBwaString.SamReadGroup}
\alias{asSamList}
\alias{asSamList.SamReadGroup}
\alias{asString}
\alias{asString.SamReadGroup}
\alias{assignBy}
\alias{assignBy.SamReadGroup}
\alias{byScanBamHeader}
\alias{byScanBamHeader.SamReadGroup}
\alias{getParameters.Bowtie2Alignment}
\alias{getParameters.BwaAlignment}
\alias{hasDescription}
\alias{hasDescription.SamReadGroup}
\alias{hasLibrary}
\alias{hasLibrary.SamReadGroup}
\alias{hasPlatform}
\alias{hasPlatform.SamReadGroup}
\alias{hasPlatformUnit}
\alias{hasPlatformUnit.SamReadGroup}
\alias{hasRunDate}
\alias{hasRunDate.SamReadGroup}
\alias{hasSample}
\alias{hasSample.SamReadGroup}
\alias{hasSequencingCenter}
\alias{hasSequencingCenter.SamReadGroup}
\alias{validate.SamReadGroup}
\alias{extractSamReadGroup}
\alias{extractSamReadGroup.FastqDataFile}
\alias{extractSamReadGroup.IlluminaFastqDataFile}
\alias{getBarcodeSequence}
\alias{getBarcodeSequence.IlluminaFastqDataFile}
\alias{getFirstSequenceInfo}
\alias{getFirstSequenceInfo.IlluminaFastqDataFile}
\alias{getFlowcellId}
\alias{getFlowcellId.IlluminaFastqDataFile}
\alias{getInstrumentId}
\alias{getInstrumentId.IlluminaFastqDataFile}
\alias{getLane}
\alias{getLane.IlluminaFastqDataFile}
\alias{getPlatform.IlluminaFastqDataFile}
\alias{getPlatformUnit}
\alias{getPlatformUnit.IlluminaFastqDataFile}
\alias{getSampleName.IlluminaFastqDataFile}
\alias{merge.SamReadGroup}
\alias{as.character.AbstractAlignment}
\alias{getAsteriskTags.AbstractAlignment}
\alias{getDefaultSamReadGroup}
\alias{getDefaultSamReadGroup.FastqDataFile}
\alias{getDefaultSamReadGroup.FastqDataSet}
\alias{getDefaultSamReadGroup.IlluminaFastqDataFile}
\alias{getFlavor.AbstractAlignment}
\alias{getFullName.AbstractAlignment}
\alias{getIndexSet.AbstractAlignment}
\alias{getInputDataSet.AbstractAlignment}
\alias{getName.AbstractAlignment}
\alias{getOptionalArguments.AbstractAlignment}
\alias{getParameterSets.AbstractAlignment}
\alias{getPath.AbstractAlignment}
\alias{getRootPath.AbstractAlignment}
\alias{getSamReadGroup}
\alias{getSamReadGroup.FastqDataFile}
\alias{getSamReadGroup.FastqDataSet}
\alias{getTags.AbstractAlignment}
\alias{isEmpty}
\alias{isEmpty.SamReadGroup}
\alias{nbrOfFiles.AbstractAlignment}
\alias{process.AbstractAlignment}
\alias{setSamReadGroup}
\alias{setSamReadGroup.FastqDataFile}
\alias{setSamReadGroup.FastqDataSet}
\alias{setTags.AbstractAlignment}
\alias{getReadGroup}
\alias{getReadGroup.BamDataFile}
\alias{hasID}
\alias{hasID.SamReadGroup}
\alias{as.character.PicardDuplicateRemoval}
\alias{getAsteriskTags.PicardDuplicateRemoval}
\alias{getFileVersion}
\alias{getFileVersion.IlluminaFastqDataFile}
\alias{getFlavor.PicardDuplicateRemoval}
\alias{getFullName.PicardDuplicateRemoval}
\alias{getInputDataSet.PicardDuplicateRemoval}
\alias{getName.PicardDuplicateRemoval}
\alias{getOptionalArguments.PicardDuplicateRemoval}
\alias{getParameterSets.PicardDuplicateRemoval}
\alias{getPath.PicardDuplicateRemoval}
\alias{getReadDirection}
\alias{getReadDirection.IlluminaFastqDataFile}
\alias{getRootPath.PicardDuplicateRemoval}
\alias{getTags.PicardDuplicateRemoval}
\alias{nbrOfFiles.PicardDuplicateRemoval}
\alias{process.PicardDuplicateRemoval}
\alias{setTags.PicardDuplicateRemoval}
\alias{as.character.BamIndexDataFile}
\alias{getIndexFile}
\alias{getIndexFile.BamDataFile}
\alias{getIndexPrefix.Bowtie2IndexFile}
\alias{getIndexStats}
\alias{getIndexStats.BamDataFile}
\alias{getReadCounts}
\alias{getReadCounts.BamDataFile}
\alias{nbrOfMappedReads}
\alias{nbrOfMappedReads.BamDataFile}
\alias{nbrOfReads.BamDataFile}
\alias{nbrOfUnmappedReads}
\alias{nbrOfUnmappedReads.BamDataFile}
\alias{extractReadStartPositions}
\alias{extractReadStartPositions.BamDataFile}
\alias{readDataFrame.BamDataFile}
\alias{readReadPositions}
\alias{readReadPositions.BamDataFile}
\alias{binTabulate.MaskedDNAString}
\alias{binTabulate}
\alias{extractRawCopyNumbers.BamDataFile}
\alias{getChipType.BamDataFile}
\alias{getChipType.BamDataSet}
\alias{getExpectedOutputFullnames.TotalCnBinnedCounting}
\alias{getFilenameExtension.BamDataFile}
\alias{getOutputFileExtension}
\alias{getOutputFileExtension.TotalCnBinnedCounting}
\alias{getOutputFileSetClass}
\alias{getOutputFileSetClass.TotalCnBinnedCounting}
\alias{getPlatform.BamDataFile}
\alias{getPlatform.BamDataSet}
\alias{smoothRawCopyNumbers}
\alias{smoothRawCopyNumbers.TotalCnBinnedCounting}
\alias{AromaUncFile}
\alias{AromaUnitNucleotideCountsFile}
\alias{allocateFromUgp}
\alias{allocateFromUgp.AromaUnitNucleotideCountsFile}
\alias{getAromaUncFile}
\alias{getAromaUncFile.AromaUgpFile}
\alias{getColumnNames.AromaUnitNucleotideCountsFile}
\alias{getExtensionPattern.AromaUnitNucleotideCountsFile}
\alias{getFilenameExtension.AromaUnitNucleotideCountsFile}
\alias{importFromBSgenome}
\alias{importFromBSgenome.AromaUnitNucleotideCountsFile}
\alias{as.character.BinnedGcNormalization}
\alias{getAsteriskTags.BinnedGcNormalization}
\alias{getGcContent}
\alias{getGcContent.AromaUnitNucleotideCountsFile}
\alias{getGcContentFile}
\alias{getGcContentFile.BinnedGcNormalization}
\alias{getOutputDataSet0.BinnedGcNormalization}
\alias{getOutputFileClass}
\alias{getOutputFileClass.BinnedGcNormalization}
\alias{getOutputFileExtension.BinnedGcNormalization}
\alias{getOutputFileSetClass.BinnedGcNormalization}
\alias{getOutputFiles.BinnedGcNormalization}
\alias{getParameters.BinnedGcNormalization}
\alias{getPath.BinnedGcNormalization}
\alias{getRootPath.BinnedGcNormalization}
\alias{process.BinnedGcNormalization}
\alias{GatkAlleleCounting}
\alias{anyDuplicated.AromaUgpFile}
\alias{getParameters.GatkAlleleCounting}
\alias{getPath.GatkAlleleCounting}
\alias{getRootPath.GatkAlleleCounting}
\alias{process.GatkAlleleCounting}
\alias{readGatkCountFile}
\alias{readGatkCountFile.GatkAlleleCounting}
\alias{writeBedDataFile}
\alias{writeBedDataFile.AromaUgpFile}
\alias{getDefaultColumnNames.AromaUnitNucleotideCountsFile}
\alias{getCombineBy}
\alias{getCombineBy.GatkAlleleCounting}
\alias{getExpectedOutputFullnames.GatkAlleleCounting}
\alias{getOutputDataSet0.GatkAlleleCounting}
\alias{getParameterSets.BwaAlignment}
\alias{getParameters.PicardDuplicateRemoval}
\alias{FastqQualityReport}
\alias{findRspReportTemplate}
\alias{findRspReportTemplate.Object}
\alias{report}
\alias{report.FastqDataFile}
\alias{report.FastqDataSet}
\alias{queryBowtie2}
\alias{getDefaultFullName.FastqDataFile}
\alias{getInputDataSet.FastqDownsampler}
\alias{getOutputPath}
\alias{getOutputPath.FastqDownsampler}
\alias{getOutputRootPath}
\alias{getOutputRootPath.FastqDownsampler}
\alias{getSampleSize}
\alias{getSampleSize.FastqDownsampler}
\alias{process.FastqDownsampler}
\alias{writeSample}
\alias{writeSample.FastqDataFile}
\alias{as.character.FastqDownsampler}
\alias{getAsteriskTags.FastqDownsampler}
\alias{getPath.FastqDownsampler}
\alias{getTags.FastqDownsampler}
\alias{as.character.FastqDataSet}
\alias{byPath.FastaReferenceSet}
\alias{byPath.FastqDataSet}
\alias{byPath.PairedEndFastqDataSet}
\alias{getDepth.FastaReferenceSet}
\alias{getFilePairs}
\alias{getFilePairs.FastqDataSet}
\alias{getParameters.TopHatAlignment}
\alias{isPaired.AbstractAlignment}
\alias{isPaired.FastqDataSet}
\alias{process.TopHatAlignment}
\alias{validate.FastaReferenceSet}
\alias{findMateFile}
\alias{findMateFile.FastqDataFile}
\alias{getMateFile}
\alias{getMateFile.FastqDataFile}
\alias{getBatchJobRegistry}
\alias{getBatchJobRegistry.default}
\alias{getBatchJobRegistryIdByMethod}
\alias{getBatchJobRegistryIdByMethod.GenericDataFileSet}
\alias{getBatchJobRegistryIdByMethod.default}
\alias{getParameters.TopHat2Alignment}
\alias{getRootPath.TopHat2Alignment}
\alias{process.TopHat2Alignment}
\alias{getPath.TopHat2Alignment}
\alias{dsApply}
\alias{getExpectedOutputPaths}
\alias{getExpectedOutputPaths.TopHat2Alignment}
\alias{getSampleNames.TopHat2Alignment}
\alias{isDone.TopHat2Alignment}
\alias{getKnownOrganisms}
\alias{getKnownOrganisms.AromaSeq}
\alias{getOrganism}
\alias{getOrganism.Arguments}
\alias{getOrganism.BamDataSet}
\alias{getOrganism.FastqDataSet}
\alias{getOrganism.SamDataSet}
\alias{skeleton}
\alias{skeleton.AromaSeq}
\alias{byName.BamDataSet}
\alias{byName.SamDataSet}
\alias{byName.FastqDataSet}
\alias{byOrganism}
\alias{byOrganism.FastaReferenceFile}
\alias{findByName.BamDataSet}
\alias{findByName.SamDataSet}
\alias{findByName.FastqDataSet}
\alias{findByOrganism}
\alias{findByOrganism.FastaReferenceFile}
\alias{agsub}
\alias{directoryItem}
\alias{directoryItem.GenericDataFile}
\alias{directoryItem.GenericDataFileSet}
\alias{directoryItems}
\alias{directoryItems.GenericDataFile}
\alias{directoryItems.GenericDataFileSet}
\alias{directoryItems.character}
\alias{directoryStructure}
\alias{directoryStructure.BamDataFile}
\alias{directoryStructure.BamDataSet}
\alias{directoryStructure.FastqDataFile}
\alias{directoryStructure.FastqDataSet}
\alias{directoryStructure.GenericDataFile}
\alias{directoryStructure.GenericDataFileSet}
\alias{directoryStructure.SamDataFile}
\alias{directoryStructure.SamDataSet}
\alias{directoryStructure.character}
\alias{directoryStructure.list}
\alias{directoryStructure<-}
\alias{directoryStructure<-.GenericDataFile}
\alias{directoryStructure<-.GenericDataFileSet}
\alias{getDataSetName}
\alias{getDataSetName.GenericDataFile}
\alias{getOrganism.AbstractAlignment}
\alias{getOrganism.AbstractIndexSet}
\alias{getOrganism.FastaReferenceFile}
\alias{getOrganismName}
\alias{getOrganismName.GenericDataFile}
\alias{getSampleName.GenericDataFile}
\alias{process.TotalCnBinnedCounting}
\alias{updateDirectoryStructure}
\alias{updateDirectoryStructure.GenericDataFileSet}
\alias{findRspReportTemplate.FastqDataFile}
\alias{as.character.AromaSeqTransform}
\alias{getAcronym}
\alias{getAcronym.AbstractAlignment}
\alias{getAcronym.AromaSeqTransform}
\alias{getAsteriskTags.AromaSeqTransform}
\alias{getFlavor.AromaSeqTransform}
\alias{getFullName.AromaSeqTransform}
\alias{getInputDataSet.AromaSeqTransform}
\alias{getName.AromaSeqTransform}
\alias{getOptionalArguments.AromaSeqTransform}
\alias{getOrganism.AromaSeqTransform}
\alias{getPath.AromaSeqTransform}
\alias{getRootPath.AromaSeqTransform}
\alias{getTags.AromaSeqTransform}
\alias{length.AromaSeqTransform}
\alias{process.AromaSeqTransform}
\alias{setTags.AromaSeqTransform}
\alias{getParameters.AromaSeqTransform}
\alias{isDone.AromaSeqTransform}
\alias{getRootPath.FastqDownsampler}
\alias{getAsteriskTags.QDNAseqEstimation}
\alias{getOutputDataSet.QDNAseqEstimation}
\alias{getRootPath.QDNAseqEstimation}
\alias{process.QDNAseqEstimation}
\alias{getBinWidth}
\alias{getBinWidth.QDNAseqEstimation}
\alias{picardValidateSamFile}
\alias{picardValidateSamFile.default}
\alias{getAsteriskTags.BamMerger}
\alias{getGroups}
\alias{getGroups.FileGroupsInterface}
\alias{getOutputDataSet.FileGroupsInterface}
\alias{getRootPath.SamTransform}
\alias{nbrOfGroups}
\alias{nbrOfGroups.FileGroupsInterface}
\alias{process.BamMerger}
\alias{validateGroups}
\alias{validateGroups.FileGroupsInterface}
\alias{validateGroups.BamMerger}
\alias{loadCacheFile}
\alias{loadCacheFile.GenericDataFile}
\alias{loadCacheFileItem}
\alias{loadCacheFileItem.GenericDataFile}
\alias{memoizedCall2}
\alias{memoizedCall2.GenericDataFile}
\alias{saveCacheFile}
\alias{saveCacheFile.GenericDataFile}
\alias{saveCacheFileItem}
\alias{saveCacheFileItem.GenericDataFile}
\alias{validateGroups.TopHat2Alignment}
\alias{assertNoDuplicated}
\alias{FileGroupsInterface}
\alias{getGroupBy}
\alias{getGroupBy.FileGroupsInterface}
\alias{getGroupNames}
\alias{getGroupNames.FileGroupsInterface}
\alias{getInputDataSet.FileGroupsInterface}
\alias{getOutputDataSet.BamMerger}
\alias{getAsteriskTags.TopHat2Alignment}
\alias{getFastaReferenceFile}
\alias{getFastaReferenceFile.AbstractIndexSet}
\alias{byOrganism.GtfDataFile}
\alias{findByOrganism.GtfDataFile}
\alias{getOrganism.GtfDataFile}
\alias{getParameters.HTSeqCounting}
\alias{getRootPath.HTSeqCounting}
\alias{getSampleNames.HTSeqCounting}
\alias{process.HTSeqCounting}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson}


\keyword{documentation}
\keyword{internal}
