%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.FastaReferenceFile}
\alias{as.character.FastqDataFile}
\alias{getCommonSeqWidth}
\alias{getCommonSeqWidth.FastqDataFile}
\alias{getGeometry}
\alias{getGeometry.FastqDataFile}
\alias{getSeqLengths}
\alias{getSeqLengths.FastaReferenceFile}
\alias{getSeqNames}
\alias{getSeqNames.FastaReferenceFile}
\alias{getTotalSeqLengths}
\alias{getTotalSeqLengths.FastaReferenceFile}
\alias{nbrOfSeqs}
\alias{nbrOfSeqs.FastaReferenceFile}
\alias{nbrOfSeqs.FastqDataFile}
\alias{readGeometry}
\alias{readGeometry.FastqDataFile}
\alias{readSeqLengths}
\alias{readSeqLengths.FastaReferenceFile}
\alias{validate.FastqDataSet}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
