%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.FastaReferenceFile}
\alias{as.character.FastqDataFile}
\alias{getCommonSeqWidth}
\alias{getCommonSeqWidth.FastqDataFile}
\alias{getGeometry}
\alias{getGeometry.FastqDataFile}
\alias{getSeqLengths}
\alias{getSeqLengths.FastaReferenceFile}
\alias{getSeqNames}
\alias{getSeqNames.FastaReferenceFile}
\alias{getTotalSeqLengths}
\alias{getTotalSeqLengths.FastaReferenceFile}
\alias{nbrOfSeqs}
\alias{nbrOfSeqs.FastaReferenceFile}
\alias{nbrOfSeqs.FastqDataFile}
\alias{readGeometry}
\alias{readGeometry.FastqDataFile}
\alias{readSeqLengths}
\alias{readSeqLengths.FastaReferenceFile}
\alias{validate.FastqDataSet}
\alias{as.character.BamDataFile}
\alias{as.character.Bowtie2Aligment}
\alias{getAsteriskTags.Bowtie2Aligment}
\alias{getFlavor}
\alias{getFlavor.Bowtie2Aligment}
\alias{getFullName.Bowtie2Aligment}
\alias{getHeader.BamDataFile}
\alias{getInputDataSet}
\alias{getInputDataSet.Bowtie2Aligment}
\alias{getName.Bowtie2Aligment}
\alias{getOutputDataSet}
\alias{getOutputDataSet.Bowtie2Aligment}
\alias{getPath.Bowtie2Aligment}
\alias{getReferenceFile}
\alias{getReferenceFile.Bowtie2Aligment}
\alias{getRootPath}
\alias{getRootPath.Bowtie2Aligment}
\alias{getTags.Bowtie2Aligment}
\alias{getTargetLengths}
\alias{getTargetLengths.BamDataFile}
\alias{getTargetNames}
\alias{getTargetNames.BamDataFile}
\alias{getTargets}
\alias{getTargets.BamDataFile}
\alias{getTotalTargetLength}
\alias{getTotalTargetLength.BamDataFile}
\alias{hasIndex}
\alias{hasIndex.BamDataFile}
\alias{isSorted}
\alias{isSorted.BamDataFile}
\alias{nbrOfFiles.Bowtie2Aligment}
\alias{nbrOfSeqs.BamDataFile}
\alias{nbrOfTargets}
\alias{nbrOfTargets.BamDataFile}
\alias{process}
\alias{process.Bowtie2Aligment}
\alias{processOne}
\alias{processOne.Bowtie2Aligment}
\alias{readHeader}
\alias{readHeader.BamDataFile}
\alias{setTags.Bowtie2Aligment}
\alias{sort.BamDataFile}
\alias{validate.BamDataSet}
\alias{bowtie2}
\alias{bowtie2.default}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
